move the various flag packages from helper to command, since command is the only package that uses them.
