+++ command/cli.go
@@ -4,9 +4,9 @@ import (
 	"flag"
 	"strings"
 
-	"github.com/hashicorp/packer/helper/enumflag"
-	kvflag "github.com/hashicorp/packer/helper/flag-kv"
-	sliceflag "github.com/hashicorp/packer/helper/flag-slice"
+	"github.com/hashicorp/packer/command/enumflag"
+	kvflag "github.com/hashicorp/packer/command/flag-kv"
+	sliceflag "github.com/hashicorp/packer/command/flag-slice"
 )
 
 //go:generate enumer -type configType -trimprefix ConfigType -transform snake
