+++ command/flag-slice/flag.go
