+++ command/flag-slice/flag_test.go
